### 2020-06-07 #2247 @kyleniemeyer
Hi @KristinaRiemer, can you take this submission on as your first editorial assignment?
