### 2020-06-08 #2247 @KristinaRiemer
@kyleniemeyer I can take this on. Can I assign myself as editor, or is that something you have to do?
