### 2020-06-08 #2247 @danielskatz
You can assign yourself, either by
```@whedon assign @KristinaRiemer as editor```
or
```@whedon assign me as editor```

After you do this, remove the waitlisted label.
