### 2020-06-11 #2247 @KristinaRiemer
@evaristor @rabernat @katmratliff @NatalieThurlby would any of you be able and willing to review this submission? Here are [the reviewing guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for reference.
