### 2020-06-12 #2328 @whedon
[REVIEW]: CollatriX: A GUI to collate MorphoMetriX outputs
**Submitting author:** @cbirdferrer (<a href="http://orcid.org/0000-0001-7763-7761">Clara Bird</a>)
**Repository:** <a href="https://github.com/cbirdferrer/collatrix" target ="_blank">https://github.com/cbirdferrer/collatrix</a>
**Version:** v1.0.5
