### 2020-06-12 #2328 @KristinaRiemer
@arfon is whedon supposed to auto-assign both reviewers as assignees to the Review issue? It looks like it missed @katmratliff here.
