### 2020-06-12 #2328 @arfon
@KristinaRiemer - Whedon can only do this once the reviewer has accepted the invite (which it looks like they now have). Whedon then updates assignments nightly so by tomorrow the assignments should look correct.
