### 2020-06-14 cbirdferrer/collatrix#2 @evaristor
Automated tests are missing
There are no automated tests. I suggest using [pytest-qt](https://pytest-qt.readthedocs.io) (or [Qt Test](https://doc.qt.io/qt-5/qttest-index.html)) in conjunction with GitHub Actions or Travis. You can find more information here:

- [PyQt: GUI Testing](https://wiki.python.org/moin/PyQt/GUI_Testing)
- [Test PyQt GUIs with QTest and unittest](http://johnnado.com/pyqt-qtest-example/)

---

- [Review for the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/2328)
