### 2020-06-14 cbirdferrer/collatrix#3 @evaristor
Community guidelines
Please add a section in the README with guidelines for third parties wishing to 1) contribute to the software, 2) report issues or problems with the software, and 3) seek support.

---

- [Review for the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/2328)
