### 2020-06-14 cbirdferrer/collatrix#4 @evaristor
The messages in the dialog boxes are too brief
The messages in the dialog boxes (titles and names of fields) are too brief. On the other hand, the description in the README (and even the comments in the code) are satisfactorily detailed. I suggest adding text boxes to the GUI with more detailed information.

---

- [Review for the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/2328)
