### 2020-06-14 cbirdferrer/collatrix#5 @evaristor
Commented-out code
Please remove the commented-out code. Lines 2 and 19 of setup.py and lines 6, 12, 16, and 17 of meta.yaml are commented out. You can find more information about why that is considered a bad idea here:

- [Please, don’t commit commented-out code](https://kentcdodds.com/blog/please-dont-commit-commented-out-code)
- [Don’t share commented-out code](https://www.nayuki.io/page/dont-share-commented-out-code)

---

- [Review for the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/2328)
