### 2020-06-14 cbirdferrer/collatrix#6 @evaristor
Build
Please show how to build the package. You could add a shell script, a Makefile, or a GitHub Actions (or Travis) configuration file that includes the command used to build the conda package.

---

- [Review for the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/2328)
