### 2020-06-14 cbirdferrer/collatrix#7 @evaristor
Output is not a valid CSV file
The header in column 1 of the output CSV file is blank. Please add a header to the first column of the output to ensure that it is a valid CSV file.

---

- [Review for the Journal of Open Source Software](https://github.com/openjournals/joss-reviews/issues/2328)
