### 2020-06-14 #2328 @evaristor
This software will be very useful to the users of MorphoMetriX since it automates the collation of the MorphoMetriX outputs with the aid of a friendly GUI. I suggest the authors make the following **Minor Revisions** before acceptance.

### From the checklist

**[Automated tests](https://github.com/cbirdferrer/collatrix/issues/2):** There are no automated tests. I suggest using [pytest-qt](https://pytest-qt.readthedocs.io) (or [Qt Test](https://doc.qt.io/qt-5/qttest-index.html)) in conjunction with GitHub Actions or Travis. You can find more information here:

- [PyQt: GUI Testing](https://wiki.python.org/moin/PyQt/GUI_Testing)
- [Test PyQt GUIs with QTest and unittest](http://johnnado.com/pyqt-qtest-example/)

**[Community guidelines](https://github.com/cbirdferrer/collatrix/issues/3):** Please add a section in the README with guidelines for third parties wishing to 1) contribute to the software, 2) report issues or problems with the software, and 3) seek support.

### Additional suggestions

**[Help messages](https://github.com/cbirdferrer/collatrix/issues/4):** The messages in the dialog boxes (titles and names of fields) are too brief. On the other hand, the description in the README (and even the comments in the code) are satisfactorily detailed. I suggest adding text boxes to the GUI with more detailed information.

**[Commented-out code](https://github.com/cbirdferrer/collatrix/issues/5):** Please remove the commented-out code. Lines [2](https://github.com/cbirdferrer/collatrix/blob/master/setup.py#L2) and [19](https://github.com/cbirdferrer/collatrix/blob/master/setup.py#L19) of `setup.py` and lines [6](https://github.com/cbirdferrer/collatrix/blob/master/meta.yaml#L6), [12](https://github.com/cbirdferrer/collatrix/blob/master/meta.yaml#L12), [16](https://github.com/cbirdferrer/collatrix/blob/master/meta.yaml#L16), and [17](https://github.com/cbirdferrer/collatrix/blob/master/meta.yaml#L17) of `meta.yaml` are commented out. You can find more information about why that is considered a bad idea here:

- [Please, don’t commit commented-out code](https://kentcdodds.com/blog/please-dont-commit-commented-out-code)
- [Don’t share commented-out code](https://www.nayuki.io/page/dont-share-commented-out-code)

**[Build](https://github.com/cbirdferrer/collatrix/issues/6):** Please show how to build the package. You could add a shell script, a Makefile, or a GitHub Actions (or Travis) configuration file that includes the command used to build the conda package.

**[Output](https://github.com/cbirdferrer/collatrix/issues/7):** The header in column 1 of the output CSV file is blank. Please add a header to the first column of the output to ensure that it is a valid CSV file.

I added these suggestions as issues on the CollatriX repo.
