### 2020-06-15 ab1f0dd
Create CODE_OF_CONDUCT.md
