+++ CODE_OF_CONDUCT.md
@@ -1,4 +1,4 @@
-# Contributor Code of Conduct
+# Contributor Covenant Code of Conduct
 
 ## Our Pledge
 
