### 2020-06-15 40c5a79
Create CONTRIBUTING.md
