### 2020-06-15 cbirdferrer/collatrix#3 @cbirdferrer
Thank you! I've added a code of conduct and contributing guidelines files to the repo and added links to those files in the README.
