### 2020-06-15 cbirdferrer/collatrix#4 @cbirdferrer
Thanks for this suggestion! We feel that adding more to the dialog box titles would be a little too wordy, which is why we added detail to the README and matched the order of the inputs to the order of the inputs in the README. That being said, I agree that the dialog could be better. To improve the link between the README and the dialog boxes, I've added numbers to the box titles that match those of the README. For example "Input 2. Do you want to use the safety?".
