### 2020-06-15 4b4866c
cleaned up code, improved dialog msgs.
