+++ collatrix/__main__.py
@@ -36,40 +36,40 @@ class App(QWidget):
 
         #do you want the Animal ID to be assigned based on the name of the folder
         items = ('yes', 'no')
-        anFold, okPressed = QInputDialog.getItem(self,"do you want the Animal ID to be assigned based on the name of the folder?", "Yes or No?",items,0,False)
+        anFold, okPressed = QInputDialog.getItem(self,"Input 1. Do you want the Animal ID to be assigned based on the name of the folder?", "Yes or No?",items,0,False)
         if okPressed and anFold:
             print("{0} Animal ID in folder name".format(anFold))
 
         #ask if they want safey net
         items = ('yes', 'no')
-        safety, okPressed = QInputDialog.getItem(self,"Safety?", "Yes or No?",items,0,False)
+        safety, okPressed = QInputDialog.getItem(self,"Input 2. Do you want to use the safety?", "Yes or No?",items,0,False)
         if okPressed and safety:
             print("{0} safety".format(safety))
         #if safety yes, ask for file
         if safety == 'yes':
             options = QFileDialog.Options()
             options |= QFileDialog.DontUseNativeDialog
-            safe_csv, _ = QFileDialog.getOpenFileName(self,"img list w/ altitudes and focal lengths", "","All Files (*);;csv files (*.csv)", options=options)
+            safe_csv, _ = QFileDialog.getOpenFileName(self,"2.1 Safety File: Image list with altitudes and other information.", "","All Files (*);;csv files (*.csv)", options=options)
             print("safety csv = {0}".format(safe_csv))
         elif safety == 'no':
             pass
 
         #animal id list?
         items = ('no','yes')
-        idchoice, okPressed = QInputDialog.getItem(self, "do you want output to only contain certain individuals?",'use animal id list?',items,0,False)
+        idchoice, okPressed = QInputDialog.getItem(self, "Input 3. Do you want output to only contain certain individuals?",'use animal id list?',items,0,False)
         if idchoice and okPressed:
             print("{0} subset list".format(idchoice))
         if idchoice == 'yes':
             options = QFileDialog.Options()
             options |= QFileDialog.DontUseNativeDialog
-            idsCSV, _ = QFileDialog.getOpenFileName(self,"file containing ID list", "","All Files (*);;csv files (*.csv)", options=options)
+            idsCSV, _ = QFileDialog.getOpenFileName(self,"3.1 File containing ID list", "","All Files (*);;csv files (*.csv)", options=options)
             if idsCSV:
                 print("ID list file = {0}".format(idsCSV))
         elif idchoice == 'no':
             pass
 
         #ask for name of output
-        outname, okPressed = QInputDialog.getText(self, "output name",'name',QLineEdit.Normal,"")
+        outname, okPressed = QInputDialog.getText(self, "Input 4. Prefix for output file.",'name',QLineEdit.Normal,"")
 
         #set up list of constants
         constants = ['Image ID', 'Image Path', 'Focal Length', 'Altitude', 'Pixel Dimension']
@@ -77,6 +77,7 @@ class App(QWidget):
         #import safety csv if safety selected
         if safety == 'yes':
             dfList = pd.read_csv(safe_csv, sep = ",")
+            dfList = dfList.dropna(how="all",axis='rows').reset_index()
             df_L = dfList.groupby('Image').first().reset_index()
             df_L['Image'] = [x.strip() for x in df_L['Image']]
         elif safety == 'no':
@@ -85,8 +86,8 @@ class App(QWidget):
         #get folders
         options = QFileDialog.Options()
         options |= QFileDialog.DontUseNativeDialog
[…]
