### 2020-06-15 cbirdferrer/collatrix#5 @cbirdferrer
Thank you for catching this! Those lines have been removed.
