### 2020-06-15 cbirdferrer/collatrix#6 @cbirdferrer
It's a short script, but I've added a shell script to the repo containing the commands I use to build the package using `conda`.
