+++ collatrix/__main__.py
@@ -18,6 +18,165 @@ import math
 from PyQt5.QtWidgets import QApplication, QWidget, QInputDialog, QLineEdit, QFileDialog, QMessageBox
 from PyQt5.QtGui import QIcon
 
+#define functions
+##duplicate check function
+def anydup(l): #we'll use this function later to check for duplicates
+    seen = set()
+    for x in l:
+        if x in seen: return True
+        seen.add(x)
+    return False
+##this is the function that does the collating
+#set up list of constants
+constants = ['Image ID', 'Image Path', 'Focal Length', 'Altitude', 'Pixel Dimension']
+def collate(csvs,measurements,nonPercMeas,df_L,safety,anFold):
+    for f in csvs: #first loop through all the csvs pulls the measurement names
+        print(f)
+        temp=pd.read_csv(f,sep='^',header=None,prefix='X',engine = 'python') #read in csv as one column
+        df00=temp.X0.str.split(',',expand=True) #split rows into columns by delimeter
+        df00 = df00.replace("",np.nan)
+        df0 = df00.dropna(how='all',axis = 'rows')
+        df0 = df0.fillna('') #replace nans by blank space
+        idx = df0.loc[df0[0] == 'Object'].index #find index (row) values of 'Object'
+        df = df0.truncate(before=idx[0]) #take subset of df starting at first row containing Object
+        head = df.iloc[0] #make list out of names in first row
+        df = df[1:] #take the data less the header row
+        df.columns = head #set the header row as the df header
+
+        wlist = df.iloc[0] #make list of all the width types
+        widths = []
+        for ww in wlist: #loop through widths row to make sure only widths are in list
+            if "Width" in str(ww):
+                widths += [str(ww)]
+
+        l = df['Object'].tolist() #make list of Object columns aka. names of measurements made
+        l = [x for x in l if pd.isna(x) == False] #eliminate all empty cells
+        l = [x for x in l if x not in constants and x != 'Object'] #elimate all other instances of Object
+        l = [x for x in l if x] #make sure no blank elements in list
+
+        measurements = measurements + l #add the measurement names to the master list
+        nonPercMeas = nonPercMeas + l #copy of the master list that does not include widths
+
+        new_header = df.columns[0:2].values.tolist() + df.iloc[0,2:].values.tolist() #merge header with width names
+        df = df[1:] #cut header row off
+        df.columns = new_header #make the updated list the new header
+
+        df = df.set_index('Object') #set Object column as the index
+        df = df.replace(r'^\s*$', np.nan, regex=True) #replace blank space with nan
+
+        if anydup(l) == True: #check for any duplicate measurement names, if exists, exit code, print error msg
+            print("please check file {0} for duplicate Objects and remove duplicates".format(f))
+            sys.exit("remove duplicate and run script again")
+        elif anydup(l) == False:
+            for i in l: #loop through list of measurement types
+                for w in (w for w in widths if w[0].isdigit()): #loop through the widths
+                    x = df.loc[i,w] #extract cell value of width of measurement type
+                    if pd.isna(x) == False: #if the cell isn't empty
+                        ww = i + "-" + w #combine the names
+                        measurements += [ww] #add this combined name to the master list
+
+    #now we're going to set up a dictionary to fill in with all the measurements
+    #that we will eventually turn into a dataframe where the keys are the columns
+    rawM = measurements
+    measurements += ['Image','Animal_ID','Altitude','Focal Length','PixD','Notes']
+    names = ['Image','Animal_ID','Altitude','Focal Length','PixD','Notes']
+    mDict = dict.fromkeys(measurements)
+    keys = list(mDict.keys())
+
[…]
