+++ requirements.txt
@@ -7,7 +7,7 @@
 
 numpy
 pandas
-pyqt=5
+pyqt==5
 setuptools
 scipy
 os
