+++ requirements.txt
@@ -7,7 +7,6 @@
 
 numpy
 pandas
-pyqt==5
 setuptools
 scipy
 os
