+++ requirements.txt
@@ -8,7 +8,4 @@
 numpy
 pandas
 setuptools
-scipy
-os
-sys
 pytest
