+++ requirements.txt
@@ -9,3 +9,4 @@ numpy
 pandas
 setuptools
 pytest
+pyqt
