+++ requirements.txt
@@ -9,4 +9,4 @@ numpy
 pandas
 setuptools
 pytest
-pyqt
+PyQt5
