+++ .github/workflows/testing.yml
@@ -6,8 +6,6 @@ name: TestCollatrix
 on:
   push:
     branches: [ master ]
-  pull_request:
-    branches: [ master ]
 
 jobs:
   build:
