### 2020-06-19 cbirdferrer/collatrix#2 @cbirdferrer
Hi! Thank you for this suggestion. We appreciate that automated testing would help improve our workflow. We have added an automated test through GitHub actions for the main function because this is the function that could be edited in the future if `MorphoMetriX` is updated. We also don’t anticipate the add-on functions being edited given that we intend for contributions to `CollatriX` to be in the form of new add-on functions.
