### 2020-06-24 cbirdferrer/collatrix#8 @katmratliff
Python version requirements
Please include the required Python version(s) in documentation.
