+++ meta.yaml
@@ -17,7 +17,7 @@ requirements:
     - python >= 3.6
     - numpy
     - pandas
-    - pyqt >= 5
+    - pyqt=5
     - setuptools
     - scipy
 
