### 2020-06-25 cbirdferrer/collatrix#8 @cbirdferrer
I've added the python version requirement to the meta.yaml and to the README. Thanks!
