### 2020-06-25 cbirdferrer/collatrix#9 @katmratliff
default input window sizes very small
The input window sizes pull up as very small on my machine (Windows 10). is there any way to change the default size so that all of the input question text is displayed? See image:
![image](https://user-images.githubusercontent.com/12630578/85729384-c437ea00-b6c6-11ea-83c3-508f35ba9554.png)

Note: submitted as part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/2328)
