### 2020-06-25 cbirdferrer/collatrix#10 @katmratliff
Add link to docs in input dialogs
Suggest adding a link to the sections of the readme with more info on the inputs if you do not want to add any more information to the dialog boxes themselves (I see your response to [this issue](https://github.com/cbirdferrer/collatrix/issues/4) but agree with @evaristor that more information would be helpful here). Perhaps you can link to the help box, which currently does not work?

![image](https://user-images.githubusercontent.com/12630578/85752974-dbcc9e00-b6d9-11ea-813e-4c44f9b57862.png)

Note: submitting as part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/2328)
