### 2020-06-25 cbirdferrer/collatrix#11 @katmratliff
testing documentation
Suggest adding in some short documentation on how to run the tests locally.

Note: submitted as part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/2328).
