### 2020-06-25 #2328 @katmratliff
Hi @KristinaRiemer, I have completed my review and have submitted the following suggestions via issues in the repository:
- https://github.com/cbirdferrer/collatrix/issues/8
- https://github.com/cbirdferrer/collatrix/issues/9
- https://github.com/cbirdferrer/collatrix/issues/10
- https://github.com/cbirdferrer/collatrix/issues/11

Under the **Documentation** checklist, I have currently left **Functionality documentation** unchecked because additional information is needed in the input dialog boxes (or a link to the documentation). I have also left **Automated tests** unchecked because there is currently no documentation on how to run the tests manually.
