### 2020-06-25 cbirdferrer/collatrix#7 @evaristor
Thank you for addressing this issue, but I'm still getting an empty header in the first column.

![image](https://user-images.githubusercontent.com/3461018/85800040-1f80d100-b6f5-11ea-882f-9b8a1c10762d.png)

You could also just remove the first column. In either case, I strongly recommend that the output be a valid CSV file.

If you believe that the problem is fixed and I'm missing something, please use an automated test to show this.
