### 2020-06-25 cbirdferrer/collatrix#2 @evaristor
Thank you for addressing my suggestion. The use of GitHub Actions to run your tests automatically is a notable improvement of your software. Congratulations!

However, one of the most relevant parts of CollatriX is the GUI and, as of now, it is not tested.

I understand the reasons you provided for not testing extensively, and I even agree that those reasons are valid for a personal project. However, for an open-source project, a code coverage of 22% is quite low.

While trying to obtain a code coverage of 100% is impractical, it is customary to aim for anything between [70% and 80%](https://www.bullseye.com/minimum.html).

Since CollatriX is open-source software, I would encourage you to obtain a code coverage of al least 40%.

---

To verify the coverage of your code, you could use the command:

```
pytest --cov-config=.coveragerc --cov=collatrix --cov-report=term-missing --verbose
```

I encourage you to replace the last line of your `.github/workflows/testing.yml` file with that command.

Since your tests are in the `collatrix/` subdirectory, you will have to add the file `.coveragerc` to the root of the project with the following content:

```
[run]
omit = collatrix/tests/*
```

Finally, don't forget to add `pytest-cov` to line 26 of `.github/workflows/testing.yml`.
