### 2020-06-25 cbirdferrer/collatrix#5 @evaristor
There is still some unnecessary text in:
- [`requirements.txt`](https://github.com/cbirdferrer/collatrix/blob/master/requirements.txt)
