### 2020-06-25 e6de8c5
updated input message box dialog
