+++ collatrix/whale_bc.py
@@ -4,7 +4,8 @@ import numpy as np
 import os, sys
 import math
 from scipy.integrate import quad
-from PyQt5.QtWidgets import QApplication, QWidget, QInputDialog, QLineEdit, QFileDialog, QMessageBox
+from PyQt5 import QtCore
+from PyQt5.QtWidgets import QApplication, QWidget, QInputDialog, QLineEdit, QFileDialog, QMessageBox, QLabel, QVBoxLayout
 from PyQt5.QtGui import QIcon
 
 class App(QWidget):
@@ -23,6 +24,13 @@ class App(QWidget):
         self.setGeometry(self.left, self.top, self.width, self.height)
         self.show()
 
+        #add message box with link to github documentation
+        msgBox = QMessageBox()
+        msgBox.setWindowTitle("For detailed input info click link below")
+        msgBox.setTextFormat(QtCore.Qt.RichText)
+        msgBox.setText('<a href = "https://github.com/cbirdferrer/collatrix">CLICK HERE</a> for detailed input instructions, \n then click on OK button to continue')
+        x = msgBox.exec_()
+
         #ask for input csv
         options = QFileDialog.Options()
         options |= QFileDialog.DontUseNativeDialog
@@ -39,21 +47,21 @@ class App(QWidget):
         message = []
         #ask if they want body Volume
         items = ('yes','no')
-        volchoice, okPressed = QInputDialog.getItem(self, 'Input 2. Do you want body volume to be calculated? (width measurements required)','',items,0,False)
+        volchoice, okPressed = QInputDialog.getItem(self, 'Input 2', 'Do you want body volume to be calculated? (width measurements required)',items,0,False)
         if okPressed and volchoice:
             print("{0} body volume calculated".format(volchoice))
 
         if volchoice == 'yes':
-            n, okPressed = QInputDialog.getText(self, "2.1 What did you name the total length measurement?","Total Length Name:", QLineEdit.Normal, "")
+            n, okPressed = QInputDialog.getText(self, "Input 2.1", "What did you name the total length measurement? \n Total Length Name:", QLineEdit.Normal, "")
             if okPressed and n != '':
                 tl_name= str(n)
-            l, okPressed = QInputDialog.getText(self, "2.2 Lower Bound","Lower Bound:", QLineEdit.Normal, "")
+            l, okPressed = QInputDialog.getText(self, "Input 2.2", "Lower Bound:", QLineEdit.Normal, "")
             if okPressed and l != '':
                 lower= int(l)
-            u, okPressed = QInputDialog.getText(self, "2.3 Upper Bound","Upper Bound:", QLineEdit.Normal, "")
+            u, okPressed = QInputDialog.getText(self, "Input 2.3", "Upper Bound:", QLineEdit.Normal, "")
             if okPressed and u != '':
                upper = int(u)
-            i, okPressed = QInputDialog.getText(self, "2.4 Interval","Interval:", QLineEdit.Normal, "")
+            i, okPressed = QInputDialog.getText(self, "Input 2.4","Interval:", QLineEdit.Normal, "")
             if okPressed and i != '':
                 interval = int(i)
             print("for body volume: length name = {0}, lower bound = {1}, upper bound = {2}, interval = {3}".format(tl_name,lower,upper,interval))
@@ -62,26 +70,26 @@ class App(QWidget):
 
         #ask if they want BAI
         items = ('yes','no')
-        baichoice, okPressed = QInputDialog.getItem(self, 'Input 3. Do you want BAI to be calculated? (you have to have measured Total_Length widths)','',items,0,False)
+        baichoice, okPressed = QInputDialog.getItem(self, 'Input 3", "Do you want BAI to be calculated? (you have to have measured Total_Length widths)',items,0,False)
[…]
