### 2020-06-25 cbirdferrer/collatrix#9 @cbirdferrer
Hi! Thank you for this suggestion, I agree the boxes were too small. I've updated the format so that the longer text is in the label and the box automatically adjusts to fit it.
