### 2020-06-26 6cb3384
added msg box with link to documentation
