### 2020-06-26 cbirdferrer/collatrix#10 @cbirdferrer
Hi! Thanks for the suggestion! I think that adding a hyperlink to the documentation within the input dialog box text is a great idea. However, after digging into the documentation and trying a few things it does not look like that is easily done. I will continue to look into ways to do this, but in the meantime I've added a message box with a hyperlink to the relevant section of the online documentation at the start of the tool.
