+++ meta.yaml
@@ -12,9 +12,9 @@ build:
 
 requirements:
   build:
-    - python >= 3.6
+    - python >=3.6
   run:
-    - python >= 3.6
+    - python >=3.6
     - numpy
     - pandas
     - pyqt=5
