### 2020-06-26 cbirdferrer/collatrix#11 @cbirdferrer
Hi! Thanks for this suggestion. I have added a section to the README with instructions on how to run pytest when the files have been downloaded locally.
