+++ .github/workflows/testing.yml
@@ -34,4 +34,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        pytest --cov-config=.coveragerc --cov=collatrix --cov-report=term-missing --verbose
+        pytest --cov=collatrix --cov-report=term-missing --verbose
