+++ .github/workflows/testing.yml
@@ -23,8 +23,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        pytest-cov
-        python -m pip install --upgrade pip
+        python -m pip install --upgrade pip pytest-cov
         pip install flake8 pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Lint with flake8
