### 2020-06-26 cbirdferrer/collatrix#7 @cbirdferrer
I apologize, I only changed one of the instances of `.to_csv` in the main function, I have double-checked that all instances (2 in the main function, 1 each in the other 2 functions) of `df.to_csv` now include an index_label: `.to_csv(outcsv,sep = ',',index_label = 'IX')`
