### 2020-06-26 cbirdferrer/collatrix#2 @cbirdferrer
Hi! I appreciate the feedback. I've added the `.coveragerc` file and updated the `testing.yml`. I've looked into using `pytest-qt` to test the GUI functions. However, I've struggled to find good documentation or examples for using `pytest-qt` and `monkey patch` on the functions I use. I was wondering if you knew of any helpful examples or resources for testing `QInputDialog.getItem`, `QInputDialog.getText`, and `QInputDialog.getOpenFileName`?
