### 2020-07-01 cbirdferrer/collatrix#2 @KristinaRiemer
@evaristor I don't know if you have any suggestions for testing these functions for @cbirdferrer?
