### 2020-07-01 cbirdferrer/collatrix#2 @evaristor
@cbirdferrer @KristinaRiemer:

[This tutorial](https://pytest-qt.readthedocs.io/en/latest/tutorial.html) looks useful since it tests [this GUI](https://github.com/nicoddemus/PySide-Examples/blob/master/examples/dialogs/findfiles.py).

You could also get inspired by these code examples that use `QInputDialog`:

- [mworion/MountWizzard4/test_tabSettImaging.py](https://github.com/mworion/MountWizzard4/blob/5047dcc96a03279ae4c837477c8678ad4d5a3b76/mw4/test/unit_tests/gui/mainWmixin/test_tabSettImaging.py)
- [PacktPublishing/Hands-On-Blockchain-for-Python-Developers/advanced_course_qt/test_button_and_dialog.py](https://github.com/PacktPublishing/Hands-On-Blockchain-for-Python-Developers/blob/2513922b658fbf41d4c75678532d970553536ccb/Chapter09/advanced_course_qt/test_button_and_dialog.py)
- [psyplot/psy-view/test_ds_widget.py](https://github.com/psyplot/psy-view/blob/2ff170aa368f51d36c7d8a6c0b97e5cb841a9909/tests/test_ds_widget.py)
- [spyder-ide/spyder/test_breakpoints.py](https://github.com/spyder-ide/spyder/blob/0a83f097eed8b1e8ac77c7eaff289a07a69f0078/spyder/plugins/editor/widgets/tests/test_breakpoints.py)
- [teampheenix/StarCraft-Casting-Tool/test_gui.py](https://github.com/teampheenix/StarCraft-Casting-Tool/blob/072cea3f9ee198132a4cd19f439b76c09a01adcf/tests/test_gui.py)
- [wonambi-python/wonambi/test_widget_notes.py](https://github.com/wonambi-python/wonambi/blob/master/tests/test_widget_notes.py)
