### 2020-07-08 cbirdferrer/collatrix#2 @cbirdferrer
Thanks @KristinaRiemer!

Hi @evaristor, thank you for finding those examples! Your comments and suggestions have helped me improve the software and I greatly appreciate that! However, I’m wondering if adding further automated testing is necessary, especially since JOSS considers well-documented manual testing to be acceptable in the review criteria- https://joss.readthedocs.io/en/latest/review_criteria.html#tests
