### 2020-07-08 cbirdferrer/collatrix#2 @evaristor
Sounds good. Go ahead and close this issue.
It looks good. 👌🏾
