### 2020-07-08 #2328 @KristinaRiemer
@evaristor thank you for confirming your acceptance! @katmratliff if you are also satisfied with this at this point, could you please confirm that?
