### 2020-07-09 #2328 @KristinaRiemer
Thank you for your reviews @evaristor and @katmratliff!

@cbirdferrer, I'm just going through the article proof linked above for correct DOIs, typos, etc., which you should do also to make sure it looks good to you.

I only found a couple minors things that you can fix when you get a chance:

- Dawson (2017) reference is missing some authors
- There is inconsistency between whether or not a comma follows "i.e."; I've been taught that it should always be "i.e.,"?
