+++ paper/paper.md
@@ -29,13 +29,13 @@ CollatriX is a graphical user interface (GUI) developed using PyQt5 to collate o
 
 # Main Features
 
-CollatriX will collate MorphoMetriX output files saved in various file structure formats (i.e., a single folder vs. across multiple folders). We also included an option to subset the collated data into a separate datasheet based on a specified list of Animal IDs provided by the user. A safety option was built in CollatriX to increase user efficiency in working through large image datasets while avoiding user input errors. For example, MorphoMetriX automatically scales length measurements in pixels to real world values (i.e. meters) from manually entered altitude, focal length, and pixel dimension values [@Torres:2020]. While this setup allows for each separate image to be scaled accordingly, there is potential for input errors when entering these values, especially when working through large datasets. CollatriX provides a safety option for users where the number of pixels in a length measurement is back calculated, and the measurement is recalculated using the correct values per image from a user provided csv.
+CollatriX will collate MorphoMetriX output files saved in various file structure formats (i.e., a single folder vs. across multiple folders). We also included an option to subset the collated data into a separate datasheet based on a specified list of Animal IDs provided by the user. A safety option was built in CollatriX to increase user efficiency in working through large image datasets while avoiding user input errors. For example, MorphoMetriX automatically scales length measurements in pixels to real world values (i.e., meters) from manually entered altitude, focal length, and pixel dimension values [@Torres:2020]. While this setup allows for each separate image to be scaled accordingly, there is potential for input errors when entering these values, especially when working through large datasets. CollatriX provides a safety option for users where the number of pixels in a length measurement is back calculated, and the measurement is recalculated using the correct values per image from a user provided csv.
 
 # Add-on Functions
 
 CollatriX also has two add-on functions for 1) calibrating altitude errors from a UAS flight and 2) calculating whale body condition (Fig. 1). The altitude calibration function follows the recommended Method 5 from @Burnett:2018, where measurements of a calibration object of known length are used to calculate the true altitude of the UAS to create a linear model used for correcting the altitude of images taken throughout the flight (for more detail see @Burnett:2018). The output of this add-on function can then be used as the safety for the main CollatriX function (Fig. 1).
 
-The whale body condition add-on function calculates two commonly used metrics of cetacean body condition. If the user used MorphoMetriX to measure perpendicular widths based on a length measurement, a common method for assessing body condition in cetaceans [@Christiansen:2016; @Dawson:2017; @Burnett:2018], the function can calculate body volume of the whale following @Christiansen:2018 and Body Area Index (BAI) following @Burnett:2018. BAI is a measure of dorsal surface area normalized by length, and CollatriX can calculate the surface area using parabolas [@Burnett:2018] or trapezoids [@Christiansen:2016]. Since MorphoMetriX allows the user to specify the number of perpendicular width segments of a length measurement, CollatriX provides the flexibility for the user to specify the number of width segments to include in the body condition calculation, i.e. 20 widths (5% increments of the total length), as well as the minimum and maximum bound in which to calculate body volume or BAI (i.e. widths between 20-85% vs. 25-80% of total length).
[…]
