+++ paper/paper.bib
@@ -23,7 +23,7 @@ doi = {10.1002/ecs2.1468}
 }
 
 @article{Dawson:2017,
-author = {{Dawson}, S.~M.},
+author = {Dawson, Stephen M. and Bowman, M. Hamish and Leunissen, Eva and Sirguey, Pascal},
 title = {{Inexpensive Aerial Photogrammetry for Studies of Whales and Large Marine Animals}},
 journal = {Frontier in Marine Science},
 year = {2017},
