+++ paper/paper.bib
@@ -11,7 +11,7 @@ doi = {10.21105/joss.01825}
 }
 
 @article{Christiansen:2016,
-author = {{Christiansen}, F. and {Dujon}, A.~M. and {Sprogis}, K.~R and {Arnould}, J.~P.~Y. and {Bejder}, L.},
+author = {Christiansen, F. and Dujon, A.~M. and Sprogis, K.~R and Arnould, J.~P.~Y. and Bejder, L.},
 title = {{Noninvasive unmanned aerial vehicle provides estimates of the energetic cost of reproduction in humpback whales}},
 journal = {Ecosphere},
 year = {2016},
