### 2020-07-09 #2328 @KristinaRiemer
That looks great @cbirdferrer!

The next steps are to: 1) [create a tagged release](https://docs.github.com/en/github/administering-a-repository/managing-releases-in-a-repository#creating-a-release) on GitHub and 2) archive the code (on Zenodo, figshare, institutional repository, etc.) to generate a DOI. If you can post a link to the DOI here when you're done please.
