### 2020-07-09 #2328 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
